Fix user controller
Fixed return of GetUserById method - it used to return Ok response even if the user was not found. Modified checks in Update method - if user wishes to change Username/Email we need to ensure they are not already taken. Also, there was a mistake in delete method - Guid cannot be null so checking for null does nothing.