Fix integration tests
The which were supposed to ensure, that testDb would be clean each time were not thread-safe. Due to async tests they would not be disposed and thus some tests failed when run simultaniously. Now DB is "created" each time but actually after it's creation each subsequent time it's data is renewed back to original.